Add kernel partition to image
The pnor will now contain a KERNEL partition, which holds the linux
kernel and initramfs that was previously placed in along with skiboot in
the PAYLOAD partition.

The PAYLOAD partition is shrunk to 1MB. Given skiboot is currently about
700KB, this should accommodate growth in the skiboot binary over time.
The remaining 15MB (21.5MB on the Habanero platform) is used for the
kernel and initramfs.

Signed-off-by: Joel Stanley <joel@jms.id.au>