Move HBI to end and increase HBI size to 25 MB
We were hitting size constraints again with the HBI section of
pnor. To get around this we will move HBI to the end and increase
the size by 8+ MB, basically filling rest of empty pnor space.
Additionally this commit moves the EECACHE section earlier in
the pnor section to avoid having to move adjust as much.