Re-apply axone layout changes and make script changed needed
Previously when we attempted to removed the MEMD section from the
axone pnor layout in order to make more room for the hostboot sections
we broke the build for configs using this layout. This commit reapplys
the layout changes and also makes chagnes to the associated scripts
to handle not having a MEMD section in the layout.