Rainier Bringup: Temporarily disable ECC injection for EECACHE
For Rainier bringup, a special SIMICS hook generates an EECACHE to workaround
not having a fully functional and fast performance SPI driver stack.  The
generated EECACHE does not have PNOR ECC on it, since SIMICS does not have a
PNOR ECC algorithm.  Once the SPI stack is good, the ECC will be added back.