p9Layouts: Add HBRT_PROXY partition
When ultravisor is enabled, some platform resources become ultravisor
privileged, consequently, the HBRT fails to access some resources (e.g.
scom read/write).

This patch adds the HBRT_PROXY (32KB) partition to the axone and default
PNOR layouts.

When ultravisor is enabled, the opal-prd will rather run the hbrt-proxy,
which is just a proxy for the HBRT that is running on a special secure
virtual machine.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>