Update makelidpkg to also generate lids for PNOR sections
For P10 on eBMC systems Hostboot will use PLDM file i/o commands
access PNOR sections. For this to work, the PNOR sections must
be lids on the BMC.  Historically the eBMC treated PNOR sections
as generic binary file. Prior to this commit the makelidpkg script
was creating .content and .header files for pnor sections we needed
to access at runtime as lids. Other scripts in the build process
will pick up these .content and .header files and assign them to
the correct .lid file on the BMC.  This commit updates the logic to
also generate .content files that can be assigned to .lid files
which Hostboot will consume IPL time. Note that the .lid files used
during IPL time will not match the .lid files used during runtime.

RTC: 208802