Simplify makelidpkg's job for IPL-time hostfw lids
Previously makelidpkg was responsible for combining secure header
binaries with their corresponding contents, aligning the data to
4KB byte boundaries, and generating empty files when needed. All
of this can be avoided by making changes to the hostboot-p10 script
genPnorImages.pl. This script can be passed a parameter which will
tell it to emit artifacts which can be used as IPL time lids. If
we pass in that parameter this all makelidpkg needs to do is copy
all files ending in .lidready into the ebmc_lids/host_fw_lids dir
and rename them with the .content in place of .lidready.