Remove <ecc/> tags from all PNOR sections except HBB
As we migrate towards representing these PNOR sections with .lid
files on the BMC we want to get the content of the .lid file to
be identical to the .bin that has historically be used for the
PNOR section. The .lid files do not have ECC, except for HBB.
ECC does not provide value anymore on the BMC as the data is no
longer stored on a physical nor chip. The secure header will inform
us if there is data integrity issues. This all being said we will
remove the ECC tags from this file so ECC will no longer be applied
to anythign except HBB.

RTC: 208802